requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
pyyaml==6.0.2

dvc==3.55.2

mlflow==2.17.2
optuna==4.0.0

fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
prometheus-client==0.21.0
python-json-logger==2.0.7

evidently==0.4.39
great-expectations==0.18.21

pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
ruff==0.7.2

matplotlib==3.9.2
seaborn==0.13.2
pathspec==0.12.1

streamlit
requests